[scratchpad] move update_batches implementation to separate file
for now it's used in benchmark only.
